Adopt pytest-mp

+++ tests/lib/fixtures/api/license.py
@@ -1,5 +1,6 @@
 from contextlib import contextmanager
 import logging
+
 import pytest
 
 
@@ -7,7 +8,7 @@ log = logging.getLogger(__name__)
 
 
 @pytest.fixture(scope='session')
-def apply_license(api):
+def apply_license(api, mp_trail, mp_message_board):
     """Create a context manager for on-the-fly license switching. The initial
     license intallation state is retained after exiting.
     """
@@ -34,29 +35,43 @@ def apply_license(api):
         """
         request = kwargs.pop('request', False)
 
-        config = api.current_version.get().config.get()
-        initial_info = config.license_info
+        instance_count = kwargs.get('instance_count', 0)
+        trial = kwargs.get('trial', False)
+        license_hash = str(license_type) + str(days) + str(instance_count) + str(trial)
+
+        initial_info_key = license_hash + '_initial_info'
 
         def teardown_license():
-            log.info('Restoring initial license.')
-            if not initial_info:
-                config.delete()
-            else:
-                initial_info['eula_accepted'] = True
-                config.post(initial_info)
+            with mp_trail(license_hash, 'finish') as finish:
+                if finish:
+                    config = api.current_version.get().config.get()
+                    initial_info = mp_message_board[initial_info_key]
+                    log.info('Restoring initial license.')
+                    if not initial_info:
+                        config.delete()
+                    else:
+                        initial_info['eula_accepted'] = True
+                        config.post(initial_info)
+                    del mp_message_board[initial_info_key]
 
         try:
-            if license_type is None:
-                log.info('Deleting current license...')
-                config.delete()
-            else:
-                log.info('Applying {} license...'.format(license_type))
-                config.install_license(license_type=license_type, days=days, **kwargs)
+            with mp_trail(license_hash, 'start') as start:
+                if start:
+                    config = api.current_version.get().config.get()
+                    mp_message_board[initial_info_key] = config.license_info
+
+                    if license_type is None:
+                        log.info('Deleting current license...')
+                        config.delete()
+                    else:
+                        log.info('Applying {} license...'.format(license_type))
+                        config.install_license(license_type=license_type, days=days, **kwargs)
             if request:
                 # We need to explictly register teardowns instead of yield-driven
                 # teardown for class_factory teardown ordering
                 request.addfinalizer(teardown_license)
-            yield config.get().license_info
+
+            yield
         finally:
             if not request:
                 teardown_license()
@@ -66,7 +81,6 @@ def apply_license(api):
 
 @pytest.fixture(scope='class')
 def no_license(apply_license, class_subrequest):
-    """Remove an active license"""
     with apply_license(None, request=class_subrequest):
         yield
